fix: Uppercase filenames' first letter
